Xcode 6.0 Beta 2 xcconfig and source files.
Turn on the new warnings in Xcode 6.0 Beta 2 (relative to Xcode 5.1).
	xcconfig/common.xcconfig

Specify the Swift optimization level, witch matches the C/C++/Objective-C optimization level for each configuration.
	xcconfig/common__config__debug.xcconfig
	xcconfig/common__config__release.xcconfig

A Swift file is required in a Target to generate a Swift Module when an Objective-C Bridging Header is specified. No existing library or target was previously using Swift so this file will be handy for using existing libraries with Swift.
	source/Empty.swift

Update the README to discuss the Empty Swift file.
	README.markdown