Update for Xcode 7.2
Set the minimum deployment target to match the OS SDK version. The mis-match may allow the application to run on a version on which it has not been tested. If an older OS version is desired, the setting can be explicitly set on a per target basis.
	xcconfig/common__target__ios.xcconfig
	xcconfig/common__target__osx.xcconfig

Turn off C++ RTTI by default to decrease binary size. Many projects have alternatives to the default implementation. The setting can be re-enabled on a per target basis, if necessary.
	xcconfig/common.xcconfig

Update the comment in the empty Swift file to clarify its usage.
	source/Empty.swift

Update the copyright date in the license.
	LICENSE

Editorial updates to the Read Me file; no significant content change.
	README.markdown